# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant monitoring SaaS platform API for website health checks, SSL monitoring, and intelligent alerting.",
        "title": "V-Insight API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/eovipmak/v-insight",
            "email": "support@v-insight.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Authenticate with email and password to receive a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to get JWT token",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged in, returns JWT token",
                        "schema": {
                            "$ref": "#/definitions/handlers.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the authenticated user's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current user info",
                "responses": {
                    "200": {
                        "description": "User information including id, email, and tenant_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new user account and create an associated tenant organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user and tenant",
                "parameters": [
                    {
                        "description": "Registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully registered, returns JWT token",
                        "schema": {
                            "$ref": "#/definitions/handlers.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or user already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get comprehensive dashboard data including stats, recent checks, and open incidents",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get complete dashboard data",
                "responses": {
                    "200": {
                        "description": "Complete dashboard data",
                        "schema": {
                            "$ref": "#/definitions/handlers.DashboardData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/dashboard/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get summary statistics for the dashboard including monitor counts, incidents, and uptime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard statistics",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "schema": {
                            "$ref": "#/definitions/handlers.DashboardStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Incident": {
            "type": "object",
            "properties": {
                "alert_rule_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "monitor_id": {
                    "type": "string"
                },
                "notified_at": {
                    "type": "string"
                },
                "resolved_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "description": "'open', 'resolved'",
                    "type": "string"
                },
                "trigger_value": {
                    "type": "string"
                }
            }
        },
        "entities.Monitor": {
            "type": "object",
            "properties": {
                "check_interval": {
                    "description": "seconds",
                    "type": "integer"
                },
                "check_ssl": {
                    "description": "whether to check SSL certificate",
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "last_checked_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ssl_alert_days": {
                    "description": "days before expiry to alert",
                    "type": "integer"
                },
                "tenant_id": {
                    "type": "integer"
                },
                "timeout": {
                    "description": "seconds",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "entities.MonitorCheck": {
            "type": "object",
            "properties": {
                "checked_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "monitor_id": {
                    "type": "string"
                },
                "response_time_ms": {
                    "type": "integer"
                },
                "ssl_expires_at": {
                    "type": "string"
                },
                "ssl_valid": {
                    "type": "boolean"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "handlers.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "handlers.DashboardData": {
            "type": "object",
            "properties": {
                "open_incidents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.IncidentWithDetails"
                    }
                },
                "recent_checks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.MonitorCheckWithMonitor"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/handlers.DashboardStats"
                }
            }
        },
        "handlers.DashboardStats": {
            "type": "object",
            "properties": {
                "average_response_time": {
                    "type": "number"
                },
                "down_count": {
                    "type": "integer"
                },
                "open_incidents": {
                    "type": "integer"
                },
                "overall_uptime": {
                    "type": "number"
                },
                "total_monitors": {
                    "type": "integer"
                },
                "up_count": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.IncidentWithDetails": {
            "type": "object",
            "properties": {
                "incident": {
                    "$ref": "#/definitions/entities.Incident"
                },
                "monitor": {
                    "$ref": "#/definitions/entities.Monitor"
                }
            }
        },
        "handlers.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handlers.MonitorCheckWithMonitor": {
            "type": "object",
            "properties": {
                "check": {
                    "$ref": "#/definitions/entities.MonitorCheck"
                },
                "monitor": {
                    "$ref": "#/definitions/entities.Monitor"
                }
            }
        },
        "handlers.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "tenant_name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "tenant_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}